Reorganization, exposing ttor through standard tasktorrent/tasktorrent.hpp, breaking down files into one file (hpp/cpp) per class
