+++ miniapp/mini_benchmarks/Makefile
@@ -1,30 +1,31 @@
 include Makefile.conf
 
 CFLAGS   += -pthread -O3 -std=c++14 -DTTOR_SHARED -DNDEBUG
-INCLUDE  += -I../../src/
+INCLUDE  += -I../../
 LIBS     += 
-LDFLAGS  +=
-SRCDIR   = ../../src
+SRCDIR   = ../../tasktorrent/src
 OBJDIR   = ../../build
 
 .PHONY: clean
 
-DEPS = $(SRCDIR)/runtime.hpp $(SRCDIR)/util.hpp $(SRCDIR)/serialization.hpp $(SRCDIR)/views.hpp $(SRCDIR)/apply_functions.hpp $(SRCDIR)/functional_extra.hpp
-OBJ  = $(OBJDIR)/serialization.o $(OBJDIR)/util.o
+DEPS := $(wildcard $(SRCDIR)/*.hpp)
+SRCS := $(wildcard $(SRCDIR)/*.cpp)
+OBJ := $(patsubst $(SRCDIR)/%.cpp,$(OBJDIR)/%.o,$(SRCS))
 
+default: all
 all: wait cholesky chain
 
 $(OBJDIR)/%.o: $(SRCDIR)/%.cpp $(DEPS)
 	$(CC) -o $@ -c $< $(CFLAGS) $(INCLUDE)
 
 wait: micro_wait.cpp $(OBJ)
-	$(CC) $(CFLAGS) -o $@ $^ $(INCLUDE) $(LIBS) $(LDFLAGS)
+	$(CC) $(CFLAGS) -o $@ $^ $(INCLUDE) $(LIBS)
 
 cholesky: micro_cholesky.cpp $(OBJ)
-	$(CC) $(CFLAGS) -o $@ $^ $(INCLUDE) $(LIBS) $(LDFLAGS)
+	$(CC) $(CFLAGS) -o $@ $^ $(INCLUDE) $(LIBS)
 
 chain: micro_deps.cpp $(OBJ)
-	$(CC) $(CFLAGS) -o $@ $^ $(INCLUDE) $(LIBS) $(LDFLAGS)
+	$(CC) $(CFLAGS) -o $@ $^ $(INCLUDE) $(LIBS)
 
 
 clean:
