+++ miniapp/mini_benchmarks/Makefile.conf.mac
@@ -1,4 +1,3 @@
 CC      = g++
 INCLUDE =
 LIBS    =
-LDFLAGS =
