+++ miniapp/mini_benchmarks/Makefile.conf.shepard
@@ -1,4 +1,3 @@
 CC      = icpc
 INCLUDE =
 LIBS    = -ltbbmalloc_proxy
-LDFLAGS =
