+++ miniapp/mini_benchmarks/Makefile.conf.travis
@@ -1,4 +1,3 @@
 CC      = g++
 INCLUDE =
 LIBS    =
-LDFLAGS =
