+++ miniapp/mini_benchmarks/micro_deps.cpp
@@ -1,5 +1,4 @@
-#include "runtime.hpp"
-#include "util.hpp"
+#include "tasktorrent/tasktorrent.hpp"
 #include <fstream>
 #include <array>
 #include <random>
