--- src/tasktorrent.hpp
@@ -1,16 +1,0 @@
-#ifndef __TTOR_TASKTORRENT_HPP__
-#define __TTOR_TASKTORRENT_HPP__
-
-#include "apply_functions.hpp"
-
-#ifndef TTOR_SHARED
-#include "communications.hpp"
-#endif
-
-#include "functional_extra.hpp"
-#include "runtime.hpp"
-#include "serialization.hpp"
-#include "util.hpp"
-#include "views.hpp"
-
-#endif
