+++ tasktorrent/src/CMakeLists.txt
@@ -2,8 +2,9 @@ enable_language(CXX)
 set(CMAKE_CXX_STANDARD 14)
 
 set(SOURCES 
-    serialization.cpp 
-    util.cpp 
+    tasks.cpp
+    threadpool_shared.cpp
+    util.cpp
 )
 
 if(TTOR_SHARED)
@@ -12,15 +13,12 @@ if(TTOR_SHARED)
     add_compile_options(-DTTOR_SHARED)
     add_library(TaskTorrent STATIC ${SOURCES})
     target_compile_options(TaskTorrent PUBLIC "-DTTOR_SHARED")
-    target_include_directories(TaskTorrent PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 else()
     # Adding communications routines
     message("Building TaskTorrent in distributed memory mode")    
-    list(APPEND SOURCES communications.cpp)# communications.hpp)
+    list(APPEND SOURCES communications.cpp threadpool_mpi.cpp activemessage.cpp message.cpp)
     add_library(TaskTorrent STATIC ${SOURCES})
-    target_include_directories(TaskTorrent PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
     find_package(MPI REQUIRED)
     target_link_libraries(TaskTorrent PUBLIC MPI::MPI_CXX)
 endif()
-
-
+target_include_directories(TaskTorrent PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../")
