+++ tasktorrent/src/message.cpp
@@ -1,0 +1,11 @@
+#ifndef TTOR_SHARED
+
+#include "message.hpp"
+
+namespace ttor {
+
+message::message(int other) : other(other){};
+
+}
+
+#endif
